# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for managing lifecycle of the tournaments",
        "title": "Tournament Management Service",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "127.0.0.1:30001",
    "basePath": "/api/v1",
    "paths": {
        "/tournaments": {
            "get": {
                "description": "get all tournaments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournaments"
                ],
                "summary": "Get Tournaments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.HostedTournamentsResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "host tournament by authorized player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournaments"
                ],
                "summary": "Host Tournament",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization info",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Tournament info",
                        "name": "tournament_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.TournamentInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.HostTournamentResponse"
                        }
                    }
                }
            }
        },
        "/tournaments/{id}": {
            "get": {
                "description": "get tournament",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournaments"
                ],
                "summary": "Get Tournament",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization info",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the tournament",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.GetTournamentResponse"
                        }
                    }
                }
            }
        },
        "/tournaments/{id}/enroll": {
            "post": {
                "description": "enroll self to the tournament",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournaments"
                ],
                "summary": "Self Enroll",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization info",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the tournament",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.SelfEnrollResponse"
                        }
                    }
                }
            }
        },
        "/tournaments/{id}/players": {
            "get": {
                "description": "get players for tournament",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournaments",
                    "players"
                ],
                "summary": "Get Players",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization info",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the tournament",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.GetPlayersResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "enroll a player that isn't a registered user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournaments",
                    "players"
                ],
                "summary": "Enroll Guest Player",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization info",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the tournament",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Guest info",
                        "name": "user_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.GuestUserInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.EnrollGuestPlayerResponse"
                        }
                    }
                }
            }
        },
        "/tournaments/{id}/players/{user_id}": {
            "put": {
                "description": "enroll a player that is a registered user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournaments",
                    "players"
                ],
                "summary": "Enroll Player",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization info",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the tournament",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the user to be added",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.EnrollPlayerResponse"
                        }
                    }
                }
            }
        },
        "/user/tournaments": {
            "get": {
                "description": "get all tournaments hosted by authorized user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournaments"
                ],
                "summary": "Hosted Tournaments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization info",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.HostedTournamentsResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "notify that user was registered",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Add User",
                "parameters": [
                    {
                        "description": "User info",
                        "name": "user_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UserInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.AddUserResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.GuestUserInfo": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                }
            }
        },
        "controllers.TournamentInfo": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "controllers.UserInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "views.AddUserResponse": {
            "type": "object"
        },
        "views.EnrollGuestPlayerResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "views.EnrollPlayerResponse": {
            "type": "object"
        },
        "views.GetPlayersResponse": {
            "type": "object",
            "properties": {
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.Player"
                    }
                }
            }
        },
        "views.GetTournamentResponse": {
            "type": "object",
            "properties": {
                "tournament": {
                    "$ref": "#/definitions/views.Tournament"
                }
            }
        },
        "views.Host": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "views.HostTournamentResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "views.HostedTournamentsResponse": {
            "type": "object",
            "properties": {
                "tournaments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.TournamentPreview"
                    }
                }
            }
        },
        "views.Player": {
            "type": "object",
            "properties": {
                "dropped": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "views.SelfEnrollResponse": {
            "type": "object"
        },
        "views.Tournament": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "host": {
                    "$ref": "#/definitions/views.Host"
                },
                "id": {
                    "type": "string"
                },
                "player_enrolled": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "total_players": {
                    "type": "integer"
                }
            }
        },
        "views.TournamentPreview": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "host": {
                    "$ref": "#/definitions/views.Host"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "total_players": {
                    "type": "integer"
                }
            }
        }
    }
}